+++ example/router/routes.js
@@ -23,7 +23,12 @@ import TimePicker from '../pages/time-picker.vue'
 import Select from '../pages/select.vue'
 import Dialog from '../pages/dialog.vue'
 import ActionSheet from '../pages/action-sheet.vue'
-import Scroll from '../pages/scroll.vue'
+import Scroll from '../pages/scroll/index.vue'
+import ScrollDefault from '../pages/scroll/default.vue'
+import ScrollConfig from '../pages/scroll/config.vue'
+import ScrollJd from '../pages/scroll/jd.vue'
+import ScrollHerizontal from '../pages/scroll/herizontal.vue'
+import ScrollToutiao from '../pages/scroll/toutiao.vue'
 import IndexList from '../pages/index-list/index.vue'
 import IndexListDefault from '../pages/index-list/default.vue'
 import IndexListCustom from '../pages/index-list/custom.vue'
@@ -150,7 +155,29 @@ const routes = [
   },
   {
     path: '/scroll',
-    component: Scroll
+    component: Scroll,
+    children: [
+      {
+        path: 'default',
+        component: ScrollDefault
+      },
+      {
+        path: 'config',
+        component: ScrollConfig
+      },
+      {
+        path: 'jd',
+        component: ScrollJd
+      },
+      {
+        path: 'herizontal',
+        component: ScrollHerizontal
+      },
+      {
+        path: 'toutiao',
+        component: ScrollToutiao
+      }
+    ]
   },
   {
     path: '/index-list',
